requirements.txt
streamlit
anthropic
Pillow
pandas
python-dotenv